qr-api/requirements.txt
Flask
Flask-CORS
qrcode[pil]
Pillow
python-barcode
gunicorn
opencv-python
numpy